Add migration 07 and update test scripts
Migration 07 adds a field "is_tested" to an answer set, along with a trigger.
This makes it easy to tell if an answer set has any answers.

Migration 07 also fixes a bug in a previous version of run_answer_set_trigger_operations(), although this bug was not exposed by any parts of the site.